% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rSynapseHelpers.R
\name{betterQuery}
\alias{betterQuery}
\title{A better query, cleans the column names}
\usage{
betterQuery(q, ...)
}
\arguments{
\item{q}{A Synapse query}
}
\description{
A better query, cleans the column names
}

